+++ socorro/processor/processor_app.py
@@ -3,10 +3,23 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-"""The processor app converts raw crashes into processed crashes."""
+"""The processor app converts raw crashes into processed crashes.
+
+It uses a fetch/transform/save model.
+
+1. fetch: a queue class determines which crash report to process next and
+   crash storage fetches the data from a crash storage source
+2. transform: a pipeline class runs the crash report through a set of
+   processing rules which results in a processed crash
+3. save: crash storage saves the data to crash storage destinations
+
+
+"""
 
 from contextlib import suppress
+from functools import partial
 import os
+import signal
 import sys
 import tempfile
 import time
@@ -25,10 +38,11 @@ from sentry_sdk.integrations.modules import ModulesIntegration
 from sentry_sdk.integrations.stdlib import StdlibIntegration
 from sentry_sdk.integrations.threading import ThreadingIntegration
 
-from socorro.app.fetch_transform_save_app import FetchTransformSaveApp
+from socorro.app.socorro_app import App
 from socorro.external.crashstorage_base import CrashIDNotFound, PolyStorageError
 from socorro.lib.libdatetime import isoformat_to_time
 from socorro.lib.libdockerflow import get_release_name
+from socorro.lib.task_manager import respond_to_SIGTERM
 
 
 CONFIG_DEFAULTS = {
@@ -92,7 +106,7 @@ def count_sentry_scrub_error(msg):
     METRICS.incr("sentry_scrub_error", 1)
 
 
-class ProcessorApp(FetchTransformSaveApp):
+class ProcessorApp(App):
     """Configman app that transforms raw crashes into processed crashes."""
 
     app_name = "processor"
@@ -101,6 +115,41 @@ class ProcessorApp(FetchTransformSaveApp):
     config_defaults = CONFIG_DEFAULTS
 
     required_config = Namespace()
+
+    # The queue class has an iterator for work items to be processed.
+    required_config.namespace("queue")
+    required_config.queue.add_option(
+        "crashqueue_class",
+        doc="an iterable that will stream work items for processing",
+        default="",
+        from_string_converter=class_converter,
+    )
+
+    # The source class has methods to fetch the data to use.
+    required_config.source = Namespace()
+    required_config.source.add_option(
+        "crashstorage_class",
+        doc="the source storage class",
+        default="socorro.external.fs.crashstorage.FSPermanentStorage",
+        from_string_converter=class_converter,
+    )
+
+    # The destination class has methods to save the transformed data to storage.
+    required_config.destination = Namespace()
+    required_config.destination.add_option(
+        "crashstorage_class",
+        doc="the destination storage class",
+        default="socorro.external.fs.crashstorage.FSPermanentStorage",
+        from_string_converter=class_converter,
+    )
+
+    required_config.producer_consumer = Namespace()
+    required_config.producer_consumer.add_option(
+        "producer_consumer_class",
+        doc="the class implements a threaded producer consumer queue",
+        default="socorro.lib.threaded_task_manager.ThreadedTaskManager",
+        from_string_converter=class_converter,
+    )
 
     # The processor is the pipeline that transforms raw crashes into
     # processed crashes.
@@ -130,6 +179,10 @@ class ProcessorApp(FetchTransformSaveApp):
         # default='socorro.processor.symbol_cache_manager.SymbolLRUCacheManager',
         from_string_converter=class_converter,
     )
+
+    def __init__(self, config):
+        super().__init__(config)
+        self.waiting_func = None
 
     @classmethod
     def configure_sentry(cls, basedir, host_id, sentry_dsn):
@@ -164,7 +217,12 @@ class ProcessorApp(FetchTransformSaveApp):
     def _basic_iterator(self):
         """Yields an infinite list of processing tasks."""
         try:
-            yield from super()._basic_iterator()
+            for x in self.queue.new_crashes():
+                if x is None or isinstance(x, tuple):
+                    yield x
+                else:
+                    yield ((x,), {})
+            yield None
         except Exception as exc:
             # NOTE(willkg): Queue code should not be throwing unhandled exceptions. If
             # it does, we should send it to sentry, log it, and then re-raise it so the
@@ -173,107 +231,138 @@ class ProcessorApp(FetchTransformSaveApp):
             self.logger.exception("error in crash id iterator")
             raise
 
-    def _transform(self, task):
-        """Runs a transform on a task.
+    def source_iterator(self):
+        """Iterate infinitely yielding tasks."""
+        while True:
+            yield from self._basic_iterator()
 
-        The ``task`` passed in is in the form of CRASHID or CRASHID:RULESET.
-        In the case of the former, it uses the default pipeline.
+    def transform(self, task, finished_func=(lambda: None)):
+        try:
+            if ":" in task:
+                crash_id, ruleset_name = task.split(":", 1)
+            else:
+                crash_id, ruleset_name = task, "default"
 
-        """
+            # Set up metrics and sentry scopes
+            with METRICS.timer("process_crash", tags=[f"ruleset:{ruleset_name}"]):
+                with sentry_sdk.push_scope() as scope:
+                    scope.set_extra("crash_id", crash_id)
+                    scope.set_extra("ruleset", ruleset_name)
 
-        if ":" in task:
-            crash_id, ruleset_name = task.split(":", 1)
-        else:
-            crash_id, ruleset_name = task, "default"
+                    # Create temporary directory context
+                    with tempfile.TemporaryDirectory(dir=self.temporary_path) as tmpdir:
+                        # Process the crash report
+                        self.process_crash(
+                            crash_id=crash_id,
+                            ruleset_name=ruleset_name,
+                            tmpdir=tmpdir,
+                        )
 
-        with METRICS.timer("process_crash", tags=[f"ruleset:{ruleset_name}"]):
-            self.process_crash(crash_id, ruleset_name)
+        finally:
+            # no matter what causes this method to end, we need to make sure
+            # that the finished_func gets called. If the new crash source is
+            # Pub/Sub, this is what removes the job from the queue.
+            try:
+                finished_func()
+            except Exception:
+                # when run in a thread, a failure here is not a problem, but if
+                # we're running all in the same thread, a failure here could
+                # derail the the whole processor. Best just log the problem
+                # so that we can continue.
+                self.logger.exception(f"Error calling finishing_func() on {task}")
 
-    def process_crash(self, crash_id, ruleset_name):
+    def process_crash(self, crash_id, ruleset_name, tmpdir):
         """Processed crash data using a specified ruleset into a processed crash.
 
-        The ``crash_id`` is a key to fetch the raw crash data from the ``source``, the
-        ``processor_class`` processes the crash and the processed crash is saved to the
-        ``destination``.
+        :arg crash_id: unique identifier for the crash report used to fetch the data and
+            save the processed data
+        :arg ruleset_name: the name of the ruleset to process the crash report with
+        :arg tmpdir: the temporary directory to use as a workspace
 
         """
-        self.logger.info("starting %s", crash_id)
-        with sentry_sdk.push_scope() as scope:
-            scope.set_extra("crash_id", crash_id)
-            scope.set_extra("ruleset", ruleset_name)
+        self.logger.info("starting %s with %s", crash_id, ruleset_name)
 
-            with tempfile.TemporaryDirectory(dir=self.temporary_path) as tmpdir:
-                self.logger.info("using tmpdir %s", tmpdir)
-                # Fetch the raw crash data
-                try:
-                    raw_crash = self.source.get_raw_crash(crash_id)
-                    dumps = self.source.get_dumps_as_files(crash_id, tmpdir)
-                except CrashIDNotFound:
-                    # If the crash isn't found, we just reject it--no need to capture
-                    # errors here
-                    self.processor.reject_raw_crash(
-                        crash_id, "crash cannot be found in raw crash storage"
-                    )
-                    return
-                except Exception as exc:
-                    sentry_sdk.capture_exception(exc)
-                    self.logger.exception("error: crash id %s: %r", crash_id, exc)
-                    self.processor.reject_raw_crash(
-                        crash_id, f"error in loading: {exc}"
-                    )
-                    return
+        self.logger.debug("fetching data %s", crash_id)
+        # Fetch crash annotations and dumps
+        try:
+            raw_crash = self.source.get_raw_crash(crash_id)
+            dumps = self.source.get_dumps_as_files(crash_id, tmpdir)
+        except CrashIDNotFound:
+            # If the crash isn't found, we just reject it--no need to capture
+            # errors here
+            self.processor.reject_raw_crash(
+                crash_id, "crash cannot be found in raw crash storage"
+            )
+            return
+        except Exception as exc:
+            sentry_sdk.capture_exception(exc)
+            self.logger.exception("error: crash id %s: %r", crash_id, exc)
+            self.processor.reject_raw_crash(crash_id, f"error in loading: {exc}")
+            return
 
-                # Fetch processed crash data--there won't be any if this crash hasn't
-                # been processed, yet
-                try:
-                    processed_crash = self.source.get_processed(crash_id)
-                    new_crash = False
-                except CrashIDNotFound:
-                    new_crash = True
-                    processed_crash = {}
+        # Fetch processed crash data--there won't be any if this crash hasn't
+        # been processed, yet
+        try:
+            processed_crash = self.source.get_processed(crash_id)
+            new_crash = False
+        except CrashIDNotFound:
+            new_crash = True
+            processed_crash = {}
 
-                # Process the crash and remove any temporary artifacts from disk
-                try:
-                    # Process the crash to generate a processed crash
-                    processed_crash = self.processor.process_crash(
-                        ruleset_name=ruleset_name,
-                        raw_crash=raw_crash,
-                        dumps=dumps,
-                        processed_crash=processed_crash,
-                        tmpdir=tmpdir,
-                    )
+        # Process the crash and remove any temporary artifacts from disk
+        # Process the crash to generate a processed crash
+        self.logger.debug("processing %s", crash_id)
+        processed_crash = self.processor.process_crash(
+            ruleset_name=ruleset_name,
+            raw_crash=raw_crash,
+            dumps=dumps,
+            processed_crash=processed_crash,
+            tmpdir=tmpdir,
+        )
 
-                    self.destination.save_processed_crash(raw_crash, processed_crash)
-                    self.logger.info("saved %s", crash_id)
-                except PolyStorageError as poly_storage_error:
-                    # Capture and log the exceptions raised by storage backends
-                    for storage_error in poly_storage_error:
-                        sentry_sdk.capture_exception(storage_error)
-                        self.logger.error(
-                            "error: crash id %s: %r", crash_id, storage_error
-                        )
-                    self.logger.warning(
-                        "error in processing or saving crash %s", crash_id
-                    )
+        # Save data to crash storage
+        try:
+            self.logger.debug("saving %s", crash_id)
+            self.destination.save_processed_crash(raw_crash, processed_crash)
+            self.logger.info("saved %s", crash_id)
+        except PolyStorageError as poly_storage_error:
+            # Capture and log the exceptions raised by storage backends
+            for storage_error in poly_storage_error:
+                sentry_sdk.capture_exception(storage_error)
+                self.logger.error("error: crash id %s: %r", crash_id, storage_error)
+            self.logger.warning("error in processing or saving crash %s", crash_id)
 
-                    # Re-raise the original exception with the correct traceback
-                    raise
+            # Re-raise the original exception with the correct traceback
+            raise
 
-            if ruleset_name == "default" and new_crash:
-                # Capture the total time for ingestion covering when the crash report was
-                # collected (submitted_timestamp) to the end of processing (now). We only
-                # want to do this for crash reports being processed for the first time.
-                collected = raw_crash.get("submitted_timestamp", None)
-                if collected:
-                    delta = time.time() - isoformat_to_time(collected)
-                    delta = delta * 1000
-                    METRICS.timing("ingestion_timing", value=delta)
+        if ruleset_name == "default" and new_crash:
+            # Capture the total time for ingestion covering when the crash report
+            # was collected (submitted_timestamp) to the end of processing (now). We
+            # only want to do this for crash reports being processed for the first
+            # time.
+            collected = raw_crash.get("submitted_timestamp", None)
+            if collected:
+                delta = time.time() - isoformat_to_time(collected)
+                delta = delta * 1000
+                METRICS.timing("ingestion_timing", value=delta)
 
         self.logger.info("completed %s", crash_id)
 
     def _setup_source_and_destination(self):
         """Instantiate classes necessary for processing."""
-        super()._setup_source_and_destination()
+        self.queue = self.config.queue.crashqueue_class(
+            self.config.queue,
+            namespace=self.app_instance_name,
+        )
+        self.source = self.config.source.crashstorage_class(
+            self.config.source,
+            namespace=self.app_name,
+        )
+        self.destination = self.config.destination.crashstorage_class(
+            self.config.destination,
+            namespace=self.app_name,
+        )
+
         if self.config.companion_process.companion_class:
             self.companion_process = self.config.companion_process.companion_class(
                 self.config.companion_process
@@ -292,17 +381,53 @@ class ProcessorApp(FetchTransformSaveApp):
         self.temporary_path = self.config.processor.temporary_path
         os.makedirs(self.temporary_path, exist_ok=True)
 
+    def _setup_task_manager(self):
+        """instantiate the threaded task manager to run the producer/consumer
+        queue that is the heart of the processor."""
+        self.logger.info("installing signal handers")
+        # Set up the signal handler for dealing with SIGTERM. The target should be this
+        # app instance so the signal handler can reach in and set the quit flag to be
+        # True. See the 'respond_to_SIGTERM' method for the more information
+        respond_to_SIGTERM_with_logging = partial(respond_to_SIGTERM, target=self)
+        signal.signal(signal.SIGTERM, respond_to_SIGTERM_with_logging)
+
+        self.task_manager = self.config.producer_consumer.producer_consumer_class(
+            self.config.producer_consumer,
+            job_source_iterator=self.source_iterator,
+            task_func=self.transform,
+        )
+
     def close(self):
         """Clean up the processor on shutdown."""
-        super().close()
-        # There is either no companion or it doesn't have a close method we can skip on
+        with suppress(AttributeError):
+            self.queue.close()
+
+        with suppress(AttributeError):
+            self.source.close()
+
+        with suppress(AttributeError):
+            self.destination.close()
+
         with suppress(AttributeError):
             self.companion_process.close()
 
-        # The processor implementation does not have a close method we can blithely skip
-        # on
         with suppress(AttributeError):
             self.processor.close()
+
+    def main(self):
+        """Main routine
+
+        Sets up the signal handlers, the source and destination crashstorage systems at
+        the theaded task manager. That starts a flock of threads that are ready to
+        shepherd tasks from the source to the destination.
+
+        """
+
+        self._setup_task_manager()
+        self._setup_source_and_destination()
+        self.task_manager.blocking_start(waiting_func=self.waiting_func)
+        self.close()
+        self.logger.info("done.")
 
 
 if __name__ == "__main__":
